Made Add Tester
Made a tester for Add, however the coverage is bad because when it attempts to make a file, I think the code I was given from this person just. Does not work? And so it causes an error which I turned into a FileNotFoundException. However because any attempts to create a file with the code appear to not work, I doubt I will be able to get my coverage up to 80%. If you comment out selective lines, however, the coverage improves although it just means not actually testing some parts.