Final Version
Reverted back to the uncommented version. This should test every method a reasonable amount. Coverage 73% due to errors in the original person's code. If you look at the coverage on the Index and Blob classes, you can see that it covers all code except for the caught parts of the try-catch blocks.